# add_data.sh: Import pre-generated boundaries PostGIS dump