from the GitHub actions on OpenDRR/boundaries#10 for speed and reliability.

Fallback to original method in case opendrr-boundaries.dump fails to download fails.

See Issue OpenDRR/boundaries#10 and PR OpenDRR/boundaries#12 for details.

----

@jvanulde Thank you for the reminder at today's afternoon meeting!
Turns out I didn't put this task into an issue and forgot about it... <grin, duck, run>

~I haven't tested it yet... (trying to meet the 4:30 p.m. PDT deadline), but will do so right away.~

Got it to run to completion after fixing my scripting errors, missing /usr/bin/time, etc.  Please see the following comment(s) for some benchmarks (and caveats).  Thanks!