@anthonyfok 2021-04-28
Initial success!  (after fixing errors and force-pushing multple times)

Some benchmarks:

opendrr-boundaries.dump is currently at about 822 MB, so its download takes roughly:

* 65 seconds at 100 Mbps
* 4m20s at 25 Mbps

The `pg_restore` run took roughly 2m5s on my computer.

For comparison, benchmark of old method (git clone with LFS and ogr2ogr):

* ~ 7 minutes at 100 Mbps for cloning (with LFS), so presumably about 28 minutes at 25 Mbps.
* ~ 5 minutes for all 9 ogr2ogr runs, 

There were some errors though:

    python-opendrr_1         | 2021-04-28T01:42:10.150205783Z pg_restore: warning: errors ignored on restore: 6
    python-opendrr_1         | 2021-04-28T01:42:10.754001788Z Command exited with non-zero status 1

so need more investigation and see whether I am running `pg_dump` (in the GitHub action at OpenDRR/boundaries) and `pg_restore` in `add_data.sh` correctly...

## Logs

The following are the raw logs which I got from (roughly) the following commands:

* `docker-compose logs -t > logfile.txt`
* `egrep '(db|python)-opendrr_1' logfile.txt | less -R` (`less -R` to show the ANSI colours)


### Before

And, for an old run using `git clone` and `ogr2ogr`:
* About 7 minutes (21:01:34 to 21:08:31) at 100 Mbps for cloning (with LFS), so presumably about 28 minutes at 25 Mbps.
* About 5 minutes (21:08:37 to 21:13:34) for all 9 ogr2ogr runs, 

```
python-opendrr_1         | 2021-03-19T21:01:34.407849526Z + git clone https://github.com/OpenDRR/boundaries.git --depth 1
python-opendrr_1         | 2021-03-19T21:01:34.411122477Z Cloning into 'boundaries'...
python-opendrr_1         | 2021-03-19T21:08:26.351510290Z Filtering content:  20% (2/10), 130.92 MiB | 5.85 MiB/s^MFiltering content:  20% (2/10), 208.43 MiB | 7.84 MiB/s^MFiltering content:  30% (3/10), 208.43 MiB | 7.84 MiB/s^MFiltering content:  30% (3/10), 293.85 MiB | 8.99 MiB/s^MFiltering content:  40% (4/10), 293.85 MiB | 8.99 MiB/s^MFiltering content:  40% (4/10), 389.39 MiB | 10.22 MiB/s^MFiltering content:  50% (5/10), 389.39 MiB | 10.22 MiB/s^MFiltering content:  50% (5/10), 447.24 MiB | 9.64 MiB/s ^MFiltering content:  60% (6/10), 447.24 MiB | 9.64 MiB/s^MFiltering content:  60% (6/10), 499.46 MiB | 10.47 MiB/s^MFiltering content:  70% (7/10), 499.46 MiB | 10.47 MiB/s^MFiltering content:  70% (7/10), 714.97 MiB | 6.68 MiB/s ^MFiltering content:  80% (8/10), 714.97 MiB | 6.68 MiB/s^MFiltering content:  80% (8/10), 1.11 GiB | 4.61 MiB/s  ^MFiltering content:  90% (9/10), 1.11 GiB | 4.61 MiB/s^MFiltering content:  90% (9/10), 1.88 GiB | 5.30 MiB/s^MFiltering content: 100% (10/10), 1.88 GiB | 5.30 MiB/s^MFiltering content: 100% (10/10), 1.88 GiB | 4.71 MiB/s, done.
python-opendrr_1         | 2021-03-19T21:08:31.011901072Z + cp model-factory/scripts/Create_MH_risk_sauid_ALL.sql ...

python-opendrr_1         | 2021-03-19T21:08:37.065068202Z  Importing Census Boundaries
python-opendrr_1         | 2021-03-19T21:08:37.065074138Z   - ogr2ogr: Importing boundaries/Geometry_ADAUID.gpkg into opendrr...
python-opendrr_1         | 2021-03-19T21:09:03.199286331Z   - ogr2ogr: Importing boundaries/Geometry_CDUID.gpkg into opendrr...
python-opendrr_1         | 2021-03-19T21:09:13.105887713Z   - ogr2ogr: Importing boundaries/Geometry_CSDUID.gpkg into opendrr...
python-opendrr_1         | 2021-03-19T21:09:31.459382703Z   - ogr2ogr: Importing boundaries/Geometry_DAUID.gpkg into opendrr...
python-opendrr_1         | 2021-03-19T21:10:18.700612079Z   - ogr2ogr: Importing boundaries/Geometry_ERUID.gpkg into opendrr...
python-opendrr_1         | 2021-03-19T21:10:36.128691930Z   - ogr2ogr: Importing boundaries/Geometry_FSAUID.gpkg into opendrr...
python-opendrr_1         | 2021-03-19T21:10:52.875484469Z   - ogr2ogr: Importing boundaries/Geometry_PRUID.gpkg into opendrr...
python-opendrr_1         | 2021-03-19T21:11:04.812921179Z   - ogr2ogr: Importing boundaries/Geometry_SAUID.gpkg into opendrr...
python-opendrr_1         | 2021-03-19T21:13:34.485719821Z   - psql: Running Update_boundaries_SAUID_table.sql...

### After

```
python-opendrr_1         | 2021-04-28T01:39:02.189832423Z  Importing Census Boundaries
python-opendrr_1         | 2021-04-28T01:39:02.189840054Z  - Trying to download pre-generated PostGIS database dump (for speed)...
python-opendrr_1         | 2021-04-28T01:39:02.197729036Z   % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
python-opendrr_1         | 2021-04-28T01:39:02.197787497Z                                  Dload  Upload   Total   Spent    Left  Speed
...
python-opendrr_1         | 2021-04-28T01:40:06.725427892Z pg_restore: connecting to database for restore
python-opendrr_1         | 2021-04-28T01:40:06.766023317Z pg_restore: dropping DATABASE PROPERTIES opendrr
python-opendrr_1         | 2021-04-28T01:40:06.766058265Z pg_restore: dropping DATABASE opendrr
python-opendrr_1         | 2021-04-28T01:40:06.766369098Z pg_restore: while PROCESSING TOC:
python-opendrr_1         | 2021-04-28T01:40:06.766381108Z pg_restore: from TOC entry 4482; 1262 19607 DATABASE opendrr postgres
python-opendrr_1         | 2021-04-28T01:40:06.766387892Z pg_restore: error: could not execute query: ERROR:  cannot drop the currently open database
python-opendrr_1         | 2021-04-28T01:40:06.766394435Z Command was: DROP DATABASE IF EXISTS opendrr;
python-opendrr_1         | 2021-04-28T01:40:06.766399933Z pg_restore: creating DATABASE "opendrr"
python-opendrr_1         | 2021-04-28T01:40:06.766789438Z pg_restore: error: could not execute query: ERROR:  database "opendrr" already exists
python-opendrr_1         | 2021-04-28T01:40:06.766800969Z Command was: CREATE DATABASE opendrr WITH TEMPLATE = template0 ENCODING = 'UTF8' LOCALE = 'en_US.utf8';
python-opendrr_1         | 2021-04-28T01:40:06.766806584Z 
python-opendrr_1         | 2021-04-28T01:40:06.766812013Z 
python-opendrr_1         | 2021-04-28T01:40:06.767081509Z pg_restore: connecting to new database "opendrr"
python-opendrr_1         | 2021-04-28T01:40:06.767093293Z pg_restore: connecting to database "opendrr" as user "postgres"
python-opendrr_1         | 2021-04-28T01:40:06.778456980Z pg_restore: creating DATABASE PROPERTIES "opendrr"
python-opendrr_1         | 2021-04-28T01:40:06.781593854Z pg_restore: connecting to new database "opendrr"
python-opendrr_1         | 2021-04-28T01:40:06.781618811Z pg_restore: connecting to database "opendrr" as user "postgres"
python-opendrr_1         | 2021-04-28T01:40:06.795157750Z pg_restore: creating SCHEMA "boundaries"
python-opendrr_1         | 2021-04-28T01:40:06.796434588Z pg_restore: from TOC entry 8; 2615 19608 SCHEMA boundaries postgres
python-opendrr_1         | 2021-04-28T01:40:06.796451745Z pg_restore: error: could not execute query: ERROR:  schema "boundaries" already exists
python-opendrr_1         | 2021-04-28T01:40:06.796476910Z Command was: CREATE SCHEMA boundaries;
python-opendrr_1         | 2021-04-28T01:40:06.796484398Z 
python-opendrr_1         | 2021-04-28T01:40:06.796490141Z 
python-opendrr_1         | 2021-04-28T01:40:06.796909969Z pg_restore: creating SCHEMA "tiger"
python-opendrr_1         | 2021-04-28T01:40:06.797493300Z pg_restore: from TOC entry 9; 2615 20780 SCHEMA tiger postgres
python-opendrr_1         | 2021-04-28T01:40:06.797508413Z pg_restore: error: could not execute query: ERROR:  schema "tiger" already exists
python-opendrr_1         | 2021-04-28T01:40:06.797515235Z Command was: CREATE SCHEMA tiger;
python-opendrr_1         | 2021-04-28T01:40:06.797521128Z 
python-opendrr_1         | 2021-04-28T01:40:06.797526563Z 
python-opendrr_1         | 2021-04-28T01:40:06.797531585Z pg_restore: creating SCHEMA "tiger_data"
python-opendrr_1         | 2021-04-28T01:40:06.797798036Z pg_restore: from TOC entry 12; 2615 21050 SCHEMA tiger_data postgres
python-opendrr_1         | 2021-04-28T01:40:06.797810047Z pg_restore: error: could not execute query: ERROR:  schema "tiger_data" already exists
python-opendrr_1         | 2021-04-28T01:40:06.797816526Z Command was: CREATE SCHEMA tiger_data;
python-opendrr_1         | 2021-04-28T01:40:06.797822271Z 
python-opendrr_1         | 2021-04-28T01:40:06.797827552Z 
python-opendrr_1         | 2021-04-28T01:40:06.798640257Z pg_restore: creating SCHEMA "topology"
python-opendrr_1         | 2021-04-28T01:40:06.798762951Z pg_restore: from TOC entry 13; 2615 20624 SCHEMA topology postgres
python-opendrr_1         | 2021-04-28T01:40:06.798816690Z pg_restore: error: could not execute query: ERROR:  schema "topology" already exists
python-opendrr_1         | 2021-04-28T01:40:06.798823402Z Command was: CREATE SCHEMA topology;
python-opendrr_1         | 2021-04-28T01:40:06.799063656Z 
python-opendrr_1         | 2021-04-28T01:40:06.799074796Z 
python-opendrr_1         | 2021-04-28T01:40:06.799080173Z pg_restore: creating COMMENT "SCHEMA topology"
python-opendrr_1         | 2021-04-28T01:40:06.802247745Z pg_restore: creating EXTENSION "fuzzystrmatch"
python-opendrr_1         | 2021-04-28T01:40:06.803521521Z pg_restore: creating COMMENT "EXTENSION fuzzystrmatch"
python-opendrr_1         | 2021-04-28T01:40:06.805640644Z pg_restore: creating EXTENSION "postgis"
python-opendrr_1         | 2021-04-28T01:40:06.805906019Z pg_restore: creating COMMENT "EXTENSION postgis"
python-opendrr_1         | 2021-04-28T01:40:06.808865165Z pg_restore: creating EXTENSION "postgis_tiger_geocoder"
python-opendrr_1         | 2021-04-28T01:40:06.809138781Z pg_restore: creating COMMENT "EXTENSION postgis_tiger_geocoder"
python-opendrr_1         | 2021-04-28T01:40:06.810947590Z pg_restore: creating EXTENSION "postgis_topology"
python-opendrr_1         | 2021-04-28T01:40:06.811254431Z pg_restore: creating COMMENT "EXTENSION postgis_topology"
python-opendrr_1         | 2021-04-28T01:40:06.812662019Z pg_restore: creating TABLE "boundaries.Geometry_ADAUID"
python-opendrr_1         | 2021-04-28T01:40:06.879100969Z pg_restore: creating SEQUENCE "boundaries.Geometry_ADAUID_fid_seq"
python-opendrr_1         | 2021-04-28T01:40:06.890489503Z pg_restore: creating SEQUENCE OWNED BY "boundaries.Geometry_ADAUID_fid_seq"
python-opendrr_1         | 2021-04-28T01:40:06.892688482Z pg_restore: creating TABLE "boundaries.Geometry_CANADA"
python-opendrr_1         | 2021-04-28T01:40:06.911292866Z pg_restore: creating SEQUENCE "boundaries.Geometry_CANADA_fid_seq"
python-opendrr_1         | 2021-04-28T01:40:06.922267943Z pg_restore: creating SEQUENCE OWNED BY "boundaries.Geometry_CANADA_fid_seq"
python-opendrr_1         | 2021-04-28T01:40:06.924921226Z pg_restore: creating TABLE "boundaries.Geometry_CDUID"
python-opendrr_1         | 2021-04-28T01:40:06.948727721Z pg_restore: creating SEQUENCE "boundaries.Geometry_CDUID_fid_seq"
python-opendrr_1         | 2021-04-28T01:40:06.959115530Z pg_restore: creating SEQUENCE OWNED BY "boundaries.Geometry_CDUID_fid_seq"
python-opendrr_1         | 2021-04-28T01:40:06.960207379Z pg_restore: creating TABLE "boundaries.Geometry_CSDUID"
python-opendrr_1         | 2021-04-28T01:40:06.979032691Z pg_restore: creating SEQUENCE "boundaries.Geometry_CSDUID_fid_seq"
python-opendrr_1         | 2021-04-28T01:40:06.989177320Z pg_restore: creating SEQUENCE OWNED BY "boundaries.Geometry_CSDUID_fid_seq"
python-opendrr_1         | 2021-04-28T01:40:06.991094004Z pg_restore: creating TABLE "boundaries.Geometry_DAUID"
python-opendrr_1         | 2021-04-28T01:40:07.014719963Z pg_restore: creating SEQUENCE "boundaries.Geometry_DAUID_fid_seq"
python-opendrr_1         | 2021-04-28T01:40:07.025070241Z pg_restore: creating SEQUENCE OWNED BY "boundaries.Geometry_DAUID_fid_seq"
python-opendrr_1         | 2021-04-28T01:40:07.026984545Z pg_restore: creating TABLE "boundaries.Geometry_ERUID"
python-opendrr_1         | 2021-04-28T01:40:07.048082758Z pg_restore: creating SEQUENCE "boundaries.Geometry_ERUID_fid_seq"
python-opendrr_1         | 2021-04-28T01:40:07.058157644Z pg_restore: creating SEQUENCE OWNED BY "boundaries.Geometry_ERUID_fid_seq"
python-opendrr_1         | 2021-04-28T01:40:07.059937481Z pg_restore: creating TABLE "boundaries.Geometry_FSAUID"
python-opendrr_1         | 2021-04-28T01:40:07.079120130Z pg_restore: creating SEQUENCE "boundaries.Geometry_FSAUID_fid_seq"
python-opendrr_1         | 2021-04-28T01:40:07.089336424Z pg_restore: creating SEQUENCE OWNED BY "boundaries.Geometry_FSAUID_fid_seq"
python-opendrr_1         | 2021-04-28T01:40:07.091316182Z pg_restore: creating TABLE "boundaries.Geometry_PRUID"
python-opendrr_1         | 2021-04-28T01:40:07.112984740Z pg_restore: creating SEQUENCE "boundaries.Geometry_PRUID_fid_seq"
python-opendrr_1         | 2021-04-28T01:40:07.124728194Z pg_restore: creating SEQUENCE OWNED BY "boundaries.Geometry_PRUID_fid_seq"
python-opendrr_1         | 2021-04-28T01:40:07.126531166Z pg_restore: creating TABLE "boundaries.Geometry_SAUID"
python-opendrr_1         | 2021-04-28T01:40:07.145309168Z pg_restore: creating SEQUENCE "boundaries.Geometry_SAUID_fid_seq"
python-opendrr_1         | 2021-04-28T01:40:07.155003062Z pg_restore: creating SEQUENCE OWNED BY "boundaries.Geometry_SAUID_fid_seq"
python-opendrr_1         | 2021-04-28T01:40:07.156751426Z pg_restore: creating DEFAULT "boundaries.Geometry_ADAUID fid"
python-opendrr_1         | 2021-04-28T01:40:07.162471478Z pg_restore: creating DEFAULT "boundaries.Geometry_CANADA fid"
python-opendrr_1         | 2021-04-28T01:40:07.164097951Z pg_restore: creating DEFAULT "boundaries.Geometry_CDUID fid"
python-opendrr_1         | 2021-04-28T01:40:07.166710120Z pg_restore: creating DEFAULT "boundaries.Geometry_CSDUID fid"
python-opendrr_1         | 2021-04-28T01:40:07.168667243Z pg_restore: creating DEFAULT "boundaries.Geometry_DAUID fid"
python-opendrr_1         | 2021-04-28T01:40:07.171279735Z pg_restore: creating DEFAULT "boundaries.Geometry_ERUID fid"
python-opendrr_1         | 2021-04-28T01:40:07.173475146Z pg_restore: creating DEFAULT "boundaries.Geometry_FSAUID fid"
python-opendrr_1         | 2021-04-28T01:40:07.175582864Z pg_restore: creating DEFAULT "boundaries.Geometry_PRUID fid"
python-opendrr_1         | 2021-04-28T01:40:07.176979435Z pg_restore: creating DEFAULT "boundaries.Geometry_SAUID fid"
python-opendrr_1         | 2021-04-28T01:40:07.179058970Z pg_restore: processing data for table "boundaries.Geometry_ADAUID"
python-opendrr_1         | 2021-04-28T01:40:13.606513490Z pg_restore: processing data for table "boundaries.Geometry_CANADA"
python-opendrr_1         | 2021-04-28T01:40:17.962165299Z pg_restore: processing data for table "boundaries.Geometry_CDUID"
python-opendrr_1         | 2021-04-28T01:40:23.976972721Z pg_restore: processing data for table "boundaries.Geometry_CSDUID"
python-opendrr_1         | 2021-04-28T01:40:35.420785241Z pg_restore: processing data for table "boundaries.Geometry_DAUID"
python-opendrr_1         | 2021-04-28T01:40:58.494412324Z pg_restore: processing data for table "boundaries.Geometry_ERUID"
python-opendrr_1         | 2021-04-28T01:41:04.748055922Z pg_restore: processing data for table "boundaries.Geometry_FSAUID"
python-opendrr_1         | 2021-04-28T01:41:13.319150476Z pg_restore: processing data for table "boundaries.Geometry_PRUID"
python-opendrr_1         | 2021-04-28T01:41:18.781306132Z pg_restore: processing data for table "boundaries.Geometry_SAUID"
python-opendrr_1         | 2021-04-28T01:41:47.822996686Z pg_restore: processing data for table "public.spatial_ref_sys"
python-opendrr_1         | 2021-04-28T01:41:47.872830804Z pg_restore: processing data for table "tiger.geocode_settings"
python-opendrr_1         | 2021-04-28T01:41:47.888251005Z pg_restore: processing data for table "tiger.pagc_gaz"
python-opendrr_1         | 2021-04-28T01:41:47.889825457Z pg_restore: processing data for table "tiger.pagc_lex"
python-opendrr_1         | 2021-04-28T01:41:47.890340800Z pg_restore: processing data for table "tiger.pagc_rules"
python-opendrr_1         | 2021-04-28T01:41:47.890840141Z pg_restore: processing data for table "topology.topology"
python-opendrr_1         | 2021-04-28T01:41:47.913347457Z pg_restore: processing data for table "topology.layer"
python-opendrr_1         | 2021-04-28T01:41:47.914074896Z pg_restore: executing SEQUENCE SET Geometry_ADAUID_fid_seq
python-opendrr_1         | 2021-04-28T01:41:47.926463134Z pg_restore: executing SEQUENCE SET Geometry_CANADA_fid_seq
python-opendrr_1         | 2021-04-28T01:41:47.931186381Z pg_restore: executing SEQUENCE SET Geometry_CDUID_fid_seq
python-opendrr_1         | 2021-04-28T01:41:47.935279079Z pg_restore: executing SEQUENCE SET Geometry_CSDUID_fid_seq
python-opendrr_1         | 2021-04-28T01:41:47.940997078Z pg_restore: executing SEQUENCE SET Geometry_DAUID_fid_seq
python-opendrr_1         | 2021-04-28T01:41:47.944044913Z pg_restore: executing SEQUENCE SET Geometry_ERUID_fid_seq
python-opendrr_1         | 2021-04-28T01:41:47.948699664Z pg_restore: executing SEQUENCE SET Geometry_FSAUID_fid_seq
python-opendrr_1         | 2021-04-28T01:41:47.954021024Z pg_restore: executing SEQUENCE SET Geometry_PRUID_fid_seq
python-opendrr_1         | 2021-04-28T01:41:47.960010484Z pg_restore: executing SEQUENCE SET Geometry_SAUID_fid_seq
python-opendrr_1         | 2021-04-28T01:41:47.965450466Z pg_restore: creating CONSTRAINT "boundaries.Geometry_ADAUID Geometry_ADAUID_pkey"
python-opendrr_1         | 2021-04-28T01:41:48.455733665Z pg_restore: creating CONSTRAINT "boundaries.Geometry_CANADA Geometry_CANADA_pkey"
python-opendrr_1         | 2021-04-28T01:41:48.610984807Z pg_restore: creating CONSTRAINT "boundaries.Geometry_CDUID Geometry_CDUID_pkey"
python-opendrr_1         | 2021-04-28T01:41:48.779971420Z pg_restore: creating CONSTRAINT "boundaries.Geometry_CSDUID Geometry_CSDUID_pkey"
python-opendrr_1         | 2021-04-28T01:41:48.957695038Z pg_restore: creating CONSTRAINT "boundaries.Geometry_DAUID Geometry_DAUID_pkey"
python-opendrr_1         | 2021-04-28T01:41:51.194131943Z pg_restore: creating CONSTRAINT "boundaries.Geometry_ERUID Geometry_ERUID_pkey"
python-opendrr_1         | 2021-04-28T01:41:51.217018296Z pg_restore: creating CONSTRAINT "boundaries.Geometry_FSAUID Geometry_FSAUID_pkey"
python-opendrr_1         | 2021-04-28T01:41:51.233548181Z pg_restore: creating CONSTRAINT "boundaries.Geometry_PRUID Geometry_PRUID_pkey"
python-opendrr_1         | 2021-04-28T01:41:51.245532952Z pg_restore: creating CONSTRAINT "boundaries.Geometry_SAUID Geometry_SAUID_pkey"
python-opendrr_1         | 2021-04-28T01:41:57.968978977Z pg_restore: creating INDEX "boundaries.Geometry_ADAUID_geom_geom_idx"
python-opendrr_1         | 2021-04-28T01:41:58.278828451Z pg_restore: creating INDEX "boundaries.Geometry_CANADA_geom_geom_idx"
python-opendrr_1         | 2021-04-28T01:41:58.839013730Z pg_restore: creating INDEX "boundaries.Geometry_CDUID_geom_geom_idx"
python-opendrr_1         | 2021-04-28T01:42:00.590521430Z pg_restore: creating INDEX "boundaries.Geometry_CSDUID_geom_geom_idx"
python-opendrr_1         | 2021-04-28T01:42:01.159791149Z pg_restore: creating INDEX "boundaries.Geometry_DAUID_geom_geom_idx"
python-opendrr_1         | 2021-04-28T01:42:03.279941259Z pg_restore: creating INDEX "boundaries.Geometry_ERUID_geom_geom_idx"
python-opendrr_1         | 2021-04-28T01:42:03.291659613Z pg_restore: creating INDEX "boundaries.Geometry_FSAUID_geom_geom_idx"
python-opendrr_1         | 2021-04-28T01:42:03.327672836Z pg_restore: creating INDEX "boundaries.Geometry_PRUID_geom_geom_idx"
python-opendrr_1         | 2021-04-28T01:42:03.340095226Z pg_restore: creating INDEX "boundaries.Geometry_SAUID_geom_geom_idx"
python-opendrr_1         | 2021-04-28T01:42:10.150205783Z pg_restore: warning: errors ignored on restore: 6
python-opendrr_1         | 2021-04-28T01:42:10.754001788Z Command exited with non-zero status 1
…